pyproject.toml
[project]
name = "marketmagicml"
version = "0.1.0"
description = "Passion project, local dependency tracking backed by github and Docker. Fun machine learning!"
authors = [
    {name = "thask",email = "tommyhaskell3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gspread (>=6.2.0,<7.0.0)",
    "oauth2client (>=4.1.3,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "newspaper3k (>=0.2.8,<0.3.0)",
    "transformers (>=4.51.2,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "lxml-html-clean (>=0.4.2,<0.5.0)",
    "newsapi-python (>=0.2.7,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
